refactor(api): split protocols into package
Bit more readable, no more giant file, better names for things
