+++ api/src/opentrons/hardware_control/protocols/stoppable.py
@@ -1,0 +1,9 @@
+from typing_extensions import Protocol
+
+
+class Stoppable(Protocol):
+    """Protocol specifying controllability of teardown"""
+
+    def clean_up(self) -> None:
+        """Get the API ready to stop cleanly."""
+        ...
