+++ robot-server/tests/conftest.py
@@ -22,7 +22,7 @@ from opentrons_shared_data.labware.dev_types import LabwareDefinition
 from opentrons import config
 from opentrons.hardware_control import (
     API,
-    BaseHardwareControl,
+    HardwareControlAPI,
     ThreadedAsyncLock,
     ThreadManager,
 )
@@ -74,7 +74,7 @@ def hardware() -> MagicMock:
 
 @pytest.fixture
 def override_hardware(hardware: MagicMock) -> None:
-    async def get_hardware_override() -> BaseHardwareControl:
+    async def get_hardware_override() -> HardwareControlAPI:
         """Override for get_hardware dependency"""
         return hardware
